recurring shell
